# packages/protocol/openapi/sync-api.yaml
openapi: 3.0.3
info:
  title: NoteApp Sync API
  version: 0.1.0
  description: >
    First-pass sync API contract for vault sync, device management, CAS commit,
    blob authorization, and ack reporting.
...
paths:
  /devices/register:
    post:
      tags: [Devices]
      summary: Register or refresh a device session
      operationId: registerDevice
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '200':
          description: Device registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterDeviceResponse'
  /devices/{deviceId}:
    delete:
      tags: [Devices]
      summary: Remove a device and revoke its future API access
      operationId: deleteDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '204':
          description: Device removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/devices:
    get:
      tags: [Devices]
      summary: List devices and per-vault ack state
      operationId: listVaultDevices
      parameters:
        - $ref: '#/components/parameters/VaultId'
      responses:
        '200':
          description: Device status list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultDeviceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/devices/heartbeat:
    post:
      tags: [Devices]
      summary: Refresh current device heartbeat for one vault
      operationId: heartbeatVaultDevice
      parameters:
        - $ref: '#/components/parameters/VaultId'
      responses:
        '200':
          description: Current device heartbeat status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultDeviceHeartbeatResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/tombstones/gc:
    post:
      tags: [Vault Sync]
      summary: Run tombstone garbage collection for one vault
      operationId: runTombstoneGc
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TombstoneGcRequest'
      responses:
        '200':
          description: Tombstone GC result and audit summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TombstoneGcResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/head:
    get:
      tags: [Vault Sync]
      summary: Get current observed remote head for a vault
      operationId: getVaultHead
      parameters:
        - $ref: '#/components/parameters/VaultId'
      responses:
        '200':
          description: Current head revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultHeadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/manifests/{revision}:
    get:
      tags: [Vault Sync]
      summary: Fetch a manifest for a specific revision
      operationId: getManifestByRevision
      parameters:
        - $ref: '#/components/parameters/VaultId'
        - $ref: '#/components/parameters/Revision'
      responses:
        '200':
          description: Manifest payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manifest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/commits:
    post:
      tags: [Vault Sync]
      summary: Submit a CAS commit for one vault revision
      operationId: createCommit
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommitRequest'
      responses:
        '200':
          description: Commit accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCommitResponse'
        '409':
          description: CAS conflict or stale base revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitConflictResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/commits/resolve-intent:
    post:
      tags: [Vault Sync]
      summary: Resolve whether a commit intent already landed remotely
      operationId: resolveCommitIntent
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveCommitIntentRequest'
      responses:
        '200':
          description: Resolution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveCommitIntentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/ack:
    post:
      tags: [Vault Sync]
      summary: Report locally applied revisions after pull success
      operationId: ackAppliedRevisions
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AckRequest'
      responses:
        '200':
          description: Ack accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/file-versions/list:
    post:
      tags: [File Versions]
      summary: List cloud-retained versions for one stable file id
      description: >
        File versions are keyed by stable file_id, not by path. A moved or
        renamed note keeps the same version chain while path_at_revision records
        where the file lived at each committed revision.
      operationId: listFileVersions
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileVersionListRequest'
      responses:
        '200':
          description: File version history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileVersionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/file-versions/{versionId}:
    patch:
      tags: [File Versions]
      summary: Update file version metadata
      description: >
        Supports user-facing metadata only: label, note, and pin state. It does
        not mutate historical blob content.
      operationId: updateFileVersion
      parameters:
        - $ref: '#/components/parameters/VaultId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileVersionUpdateRequest'
      responses:
        '200':
          description: Updated file version metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileVersionUpdateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/blobs/check:
    post:
      tags: [Blobs]
      summary: Check whether blobs already exist remotely
      operationId: checkBlobs
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlobCheckRequest'
      responses:
        '200':
          description: Blob existence result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobCheckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/blobs/upload-init:
    post:
      tags: [Blobs]
      summary: Issue short-lived upload capability for encrypted blobs
      operationId: initBlobUpload
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInitRequest'
      responses:
        '200':
          description: Upload capability issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadInitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/blobs/resumable-upload-init:
    post:
      tags: [Blobs]
      summary: Issue resumable upload capability and current chunk state
      operationId: initResumableBlobUpload
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumableUploadInitRequest'
      responses:
        '200':
          description: Resumable upload capability issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumableUploadInitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/blobs/resumable-upload-complete:
    post:
      tags: [Blobs]
      summary: Finalize resumable blob uploads after all chunks are present
      operationId: completeResumableBlobUpload
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumableUploadCompleteRequest'
      responses:
        '200':
          description: Final upload verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumableUploadCompleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/blobs/download-init:
    post:
      tags: [Blobs]
      summary: Issue short-lived download capability for encrypted blobs
      operationId: initBlobDownload
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadInitRequest'
      responses:
        '200':
          description: Download capability issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadInitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vaults/{vaultId}/blobs/resumable-download-init:
    post:
      tags: [Blobs]
      summary: Issue download capability for explicit encrypted byte ranges
      operationId: initResumableBlobDownload
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumableDownloadInitRequest'
      responses:
        '200':
          description: Ranged download capability issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumableDownloadInitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    VaultId:
      name: vaultId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    Revision:
      name: revision
      in: path
      required: true
      schema:
        type: integer
        minimum: 0
...
  responses:
    Unauthorized:
      description: Authentication required or token revoked
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not allowed for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterDeviceRequest:
      type: object
      additionalProperties: false
      required:
        - device_name
        - platform
      properties:
        device_name:
          type: string
          minLength: 1
        platform:
          type: string
          enum: [desktop, mobile, web]
        app_version:
          type: string
        protocol_version:
          type: string
          default: v1
    RegisterDeviceResponse:
      type: object
      additionalProperties: false
      required:
        - device_id
        - access_token
      properties:
        device_id:
          type: string
        access_token:
          type: string
        expires_at:
          type: string
          format: date-time
...
    VaultDeviceListResponse:
      type: object
      additionalProperties: false
      required:
        - vault_id
        - head_revision
        - inactive_after_ms
        - devices
      properties:
        vault_id:
          type: string
          minLength: 1
        head_revision:
          type: integer
          minimum: 0
        inactive_after_ms:
          type: integer
          minimum: 1
        devices:
          type: array
          items:
            $ref: '#/components/schemas/VaultDeviceRecord'
    VaultDeviceHeartbeatResponse:
      type: object
      additionalProperties: false
      required:
        - vault_id
        - device_id
        - last_seen_at_ms
        - acked_revision
        - head_revision
      properties:
        vault_id:
          type: string
          minLength: 1
        device_id:
          type: string
          minLength: 1
        last_seen_at_ms:
          type: integer
          minimum: 0
        acked_revision:
          type: integer
          minimum: 0
        head_revision:
          type: integer
          minimum: 0
    TombstoneGcRequest:
      type: object
      additionalProperties: false
      properties:
        now_ms:
          type: integer
          minimum: 0
        min_retention_ms:
          type: integer
          minimum: 0
        inactive_after_ms:
          type: integer
          minimum: 1
...
    TombstoneGcResponse:
      type: object
      additionalProperties: false
      required:
        - vault_id
        - run_id
        - ran_at_ms
        - base_revision
        - head_revision
        - active_device_ids
        - active_device_count
        - min_retention_ms
        - inactive_after_ms
        - reclaimed_tombstones
        - blocked_tombstones
        - reclaimed_count
        - reason
      properties:
        vault_id:
          type: string
          minLength: 1
        run_id:
          type: string
          minLength: 1
        ran_at_ms:
          type: integer
          minimum: 0
        base_revision:
          type: integer
          minimum: 0
        new_revision:
          oneOf:
            - type: integer
              minimum: 1
            - type: 'null'
        head_revision:
          type: integer
          minimum: 0
        active_device_ids:
          type: array
          items:
            type: string
            minLength: 1
        active_device_count:
          type: integer
          minimum: 0
        min_retention_ms:
          type: integer
          minimum: 0
        inactive_after_ms:
          type: integer
          minimum: 1
        reclaimed_tombstones:
          type: array
          items:
            $ref: '#/components/schemas/TombstoneGcTombstone'
        blocked_tombstones:
          type: array
          items:
            $ref: '#/components/schemas/TombstoneGcBlockedTombstone'
        reclaimed_count:
          type: integer
          minimum: 0
        reason:
          type: string
          enum:
            - no_head_manifest
            - reclaimed
            - no_reclaimable_tombstones
    VaultHeadResponse:
      type: object
      additionalProperties: false
      required:
        - vault_id
        - head_revision
        - manifest_summary
      properties:
        vault_id:
          type: string
        head_revision:
          type: integer
          minimum: 0
        manifest_summary:
          oneOf:
            - type: string
              minLength: 1
            - type: 'null'
    CreateCommitRequest:
      type: object
      additionalProperties: false
      required:
        - commit_intent_id
        - base_revision
        - intent_manifest_hash
        - manifest
      properties:
        commit_intent_id:
          type: string
          minLength: 1
        base_revision:
          type: integer
          minimum: 0
          description: Must exactly match manifest.base_revision.
        created_by_device:
          type: string
        intent_manifest_hash:
          type: string
          minLength: 1
        intent_delete_seq_upper_bound:
          type: integer
          minimum: 1
        manifest:
          $ref: '#/components/schemas/Manifest'
        blob_refs:
          type: array
          items:
            $ref: '#/components/schemas/BlobRef'
        file_version_directives:
          type: array
          description: >
            Optional per-file version metadata for this commit. Use
            manual_meeting_checkpoint for an explicit meeting version. Autosave
            drafts must not submit this field or create cloud versions.
          items:
            $ref: '#/components/schemas/FileVersionCommitDirective'
    CreateCommitResponse:
      type: object
      additionalProperties: false
      required:
        - vault_id
        - new_revision
        - head_manifest_summary
        - acked_revision_for_device
      properties:
        vault_id:
          type: string
        new_revision:
          type: integer
          minimum: 1
        head_manifest_summary:
          type: string
          minLength: 1
        acked_revision_for_device:
          type: integer
          minimum: 1
    CommitConflictResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - current_head_revision
        - current_manifest_summary
      properties:
        code:
          type: string
          enum: [base_revision_conflict, manifest_conflict]
        current_head_revision:
          type: integer
          minimum: 0
        current_manifest_summary:
          oneOf:
            - type: string
              minLength: 1
            - type: 'null'
...
    Manifest:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - vault_id
        - revision
        - base_revision
        - created_by_device
        - created_at
        - files
        - tombstones
        - summary_hash
      properties:
        schema_version:
          type: string
          enum: [v1]
        vault_id:
          type: string
          minLength: 1
        revision:
          type: integer
          minimum: 0
        base_revision:
          type: integer
          minimum: 0
        created_by_device:
          type: string
          minLength: 1
        created_at:
          type: integer
          minimum: 0
        files:
          type: array
          items:
            $ref: '#/components/schemas/ManifestFileEntry'
        tombstones:
          type: array
          items:
            $ref: '#/components/schemas/Tombstone'
        summary_hash:
          type: string
          minLength: 1
...
security:
  - bearerAuth: []